+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - `Session` is a singleton.
 - Updated major dependencies.
 - Minimal node.js version is 10.
+- Removed `Server.get` method.
 
 ## 0.17.3
 ### Updates